Remove artist relations and musicbrainz json release dump
This data has not been used in Spark for a long-long time so removing
it before doing rest of code cleanup.
