+++ listenbrainz/spark/background.py
@@ -13,7 +13,6 @@ from listenbrainz.spark.handlers import (
     handle_model,
     handle_recommendations,
     handle_sitewide_entity,
-    notify_artist_relation_import,
     notify_mapping_import,
     handle_missing_musicbrainz_data,
     cf_recording_recommendations_complete,
@@ -154,7 +153,6 @@ class BackgroundJobProcessor:
             "cf_recommendations_recording_candidate_sets": handle_candidate_sets,
             "cf_recommendations_recording_recommendations": handle_recommendations,
             "import_mapping": notify_mapping_import,
-            "import_artist_relation": notify_artist_relation_import,
             "missing_musicbrainz_data": handle_missing_musicbrainz_data,
             "cf_recommendations_recording_mail": cf_recording_recommendations_complete,
             "similar_users": handle_similar_users,
