+++ listenbrainz_spark/ftp/download.py
@@ -1,16 +1,8 @@
 import logging
-import os
-import tempfile
 import time
 
-from listenbrainz_spark import config
-from listenbrainz_spark.dump import DumpType, ListensDump
-from listenbrainz_spark.exceptions import DumpNotFoundException
+from listenbrainz_spark.dump import DumpType
 from listenbrainz_spark.ftp import ListenBrainzFTPDownloader
-
-# mbid_msid_mapping_with_matchable is used.
-# refer to: http://ftp.musicbrainz.org/pub/musicbrainz/listenbrainz/labs/mappings/
-ARTIST_RELATION_DUMP_ID_POS = 5
 
 FULL = 'full'
 INCREMENTAL = 'incremental'
@@ -52,46 +44,3 @@ class ListenbrainzDataDownloader(ListenBrainzFTPDownloader):
 
     def load_listens(self, directory, listens_dump_id=None, dump_type: DumpType = DumpType.FULL) -> (str, str, int):
         return self.download_listens(directory, listens_dump_id, dump_type)
-
-    def download_artist_relation(self, directory, artist_relation_dump_id=None):
-        """ Download artist relation to dir passed as an argument.
-
-            Args:
-                directory (str): Dir to save artist relation locally.
-                artist_relation_dump_id (int): Unique identifier of artist relation to be downloaded.
-                    If not provided, most recent artist relation will be downloaded.
-
-            Returns:
-                dest_path (str): Local path where artist relation has been downloaded.
-                artist_relation_file_name (str): file name of downloaded artist relation.
-        """
-        self.connection.cwd(config.FTP_ARTIST_RELATION_DIR)
-        dump = self.list_dir()
-        req_dump = self.get_dump_name_to_download(dump, artist_relation_dump_id, ARTIST_RELATION_DUMP_ID_POS)
-
-        self.connection.cwd(req_dump)
-        artist_relation_file_name = req_dump + '.tar.bz2'
-
-        t0 = time.monotonic()
-        logger.info('Downloading {} from FTP...'.format(artist_relation_file_name))
-        dest_path = self.download_dump(artist_relation_file_name, directory)
-        logger.info('Done. Total time: {:.2f} sec'.format(time.monotonic() - t0))
-
-        return dest_path, artist_relation_file_name
-
-    def download_release_json_dump(self, directory):
-        self.connection.cwd(config.FTP_MUSICBRAINZ_DIR)
-        with tempfile.TemporaryDirectory() as temp_dir:
-            temp_file = os.path.join(temp_dir, "LATEST")
-            self.download_file_binary("LATEST", temp_file)
-            with open(temp_file) as f:
-                dump_name = f.readline().strip()
-        self.connection.cwd(dump_name)
-
-        filename = "release.tar.xz"
-        logger.info(f"Downloading {filename} of dump {dump_name} from FTP...")
-        t0 = time.monotonic()
-        dest = os.path.join(directory, filename)
-        self.download_file_binary(filename, dest)
-        logger.info(f"Done. Total time: {time.monotonic() - t0:.2f} sec")
-        return dest
