+++ listenbrainz_spark/schema.py
@@ -97,15 +97,6 @@ model_metadata_schema = [
 ]
 
 
-artist_relation_schema = [
-    StructField('id_0', IntegerType(), nullable=False), # artist credit
-    StructField('name_1', StringType(), nullable=False), # artist name
-    StructField('name_0', StringType(), nullable=False),
-    StructField('id_1', IntegerType(), nullable=False),
-    StructField('score', FloatType(), nullable=False),
-]
-
-
 dataframe_metadata_schema = [
     StructField('dataframe_created', TimestampType(), nullable=False),  # Timestamp when dataframes are created and saved in HDFS.
     StructField('dataframe_id', StringType(), nullable=False),  # dataframe id or identification string of dataframe.
@@ -132,7 +123,6 @@ import_metadata_schema = [
 # Although, we try to keep it sorted in the actual definition itself, we
 # also sort it programmatically just in case
 model_metadata_schema = StructType(sorted(model_metadata_schema, key=lambda field: field.name))
-artist_relation_schema = StructType(sorted(artist_relation_schema, key=lambda field: field.name))
 dataframe_metadata_schema = StructType(sorted(dataframe_metadata_schema, key=lambda field: field.name))
 import_metadata_schema = StructType(sorted(import_metadata_schema, key=lambda field: field.name))
