--- relations/config.py.sample
@@ -1,7 +1,0 @@
-#
-# Database connection.
-
-# If you are connecting to a musicbrainz-docker instance for MusicBrainz data, this string should work:
-# DB_CONNECT = "dbname=musicbrainz_db user=musicbrainz host=localhost port=5432 password=musicbrainz"
-
-DB_CONNECT = ""
