--- relations/data_dump_files/README_data_dump_files.md
@@ -1,1 +1,0 @@
-Please note -- the files in this directory are here for inclusion in data dump files. They do not apply to the code in this repository!
