--- relations/requirements.txt
@@ -1,4 +1,0 @@
-click==7.1.2
-psycopg2-binary==2.8.6
-ujson==5.4.0
-nose==1.3.7
