--- relations/test.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-nosetests tests/test_artist_relations.py
